requirements/base.txt
Cerberus==1.3.2
Django==3.0.8
django-cors-headers==3.4.0
django-environ==0.4.5
gunicorn==20.0.4
factory-boy==2.11.1
openpyxl==2.6.2
Pillow==7.2.0
psycopg2-binary==2.8.5
python-magic==0.4.18
requests==2.24.0